Move normalizeDirection to Direction itself
