+++ packages/game-server/src/shared/physics.ts
@@ -1,5 +1,3 @@
-import { Direction } from "./direction";
-
 export type Vector2 = { x: number; y: number };
 
 export function roundVector2(vector: Vector2): Vector2 {
@@ -17,33 +15,4 @@ export function normalizeVector(vector: Vector2): Vector2 {
     x: vector.x / magnitude,
     y: vector.y / magnitude,
   };
-}
-
-export function normalizeDirection(direction: Direction): Vector2 {
-  switch (direction) {
-    case Direction.Down: {
-      return { x: 0, y: -1 };
-    }
-    case Direction.DownLeft: {
-      return { x: -1, y: -1 };
-    }
-    case Direction.DownRight: {
-      return { x: 1, y: -1 };
-    }
-    case Direction.Left: {
-      return { x: -1, y: 0 };
-    }
-    case Direction.Right: {
-      return { x: 1, y: 0 };
-    }
-    case Direction.Up: {
-      return { x: 0, y: 1 };
-    }
-    case Direction.UpLeft: {
-      return { x: -1, y: 1 };
-    }
-    case Direction.UpRight: {
-      return { x: 1, y: 1 };
-    }
-  }
 }
